backend/requirements.txt
Django
djangorestframework
psycopg2-binary
django-cors-headers
pandas
httpx
logfire
pydantic-ai
fastapi
pydantic_settings
sqlmodel